+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.10.0 - 2024-07-14
+
 ### Added
 
 - Configurable parallel jobs when running `qft evaluate-compression`